added controllers till settings menu
